Fix: fix DB not initialized
